Remove doubles.  initDegutil() in degutil.c causes the game to crash when Quit is selected if converted to floats, so leaving this one alone for now.  The remaining doubles can't be removed until libBulletML has doubles removed.

+++ noiz2sa/src/vector.c
@@ -67,7 +67,7 @@ int vctCheckSide(Vector *checkPos, Vector *pos1, Vector *pos2) {
 }
 
 int vctSize(Vector *v) {
-  return sqrt(v->x*v->x + v->y*v->y);
+  return sqrtf(v->x*v->x + v->y*v->y);
 }
 
 int vctDist(Vector *v1, Vector *v2) {
